+++ backend/queue/logger.go
@@ -1,7 +1,0 @@
-package queue
-
-import "github.com/sirupsen/logrus"
-
-var logger = logrus.WithFields(logrus.Fields{
-	"component": "queue",
-})