+++ CHANGELOG.md
@@ -37,6 +37,7 @@ the older versions are now deprecated.
 - Web ui entity recent events are sorted by last ok
 
 ### Fixed
+- Fixed a bug where gRPC would log non-json messages directly to stderr.
 - Fixes a bug in `sensuctl cluster health` so the correct error is handled.
 - Fixed a bug where assets could not extract git tarballs.
 - Fixed a bug where assets would not install if given cache directory was a